Fix tests

+++ src/main/java/dev/muscaw/monitor/weather/domain/Temperature.java
@@ -1,3 +1,3 @@
 package dev.muscaw.monitor.weather.domain;
 
-public record Temperature(float temperatureC) {}
+public record Temperature(double temperatureC) {}
