+++ src/main/java/dev/muscaw/monitor/weather/domain/Wind.java
@@ -1,3 +1,3 @@
 package dev.muscaw.monitor.weather.domain;
 
-public record Wind(float windKph, float gustKph, CardinalDirection direction) {}
+public record Wind(double windKph, double gustKph, CardinalDirection direction) {}
